Fixed that crash related with arresting cells
When trying to select a cell and failing, the cell selected flag was not
being set to false. This caused a various amount of unintended stuff
happening but it was all hard to notice except when an arrested cell was
involved which crashed the program.
example:
select a cell > arrest it > when you lift the finger that cell remains
selected > you tap next to it and nothing should happen, instead because
of that flag being true you call newCell() with cellState::arrested > try
to init a cell as arrested > the switch has no arrested case cause it's
not supposed to happen > you don't initialise a texture for the new cell
and when you call gldrawarrays opengl gets rekt. ggwp